admin dashboard completed

+++ tntcli/src/components/todo/CompletedTodoList.js
@@ -6,7 +6,8 @@ import { deleteTodo } from "../../actions/todoAction";
 
 class CompletedTodoList extends Component {
   onDeleteClick = (userId, taskId) => {
-    this.props.deleteTodo(userId, taskId);
+    window.confirm("Are you sure you want to delete the TODO?") &&
+      this.props.deleteTodo(userId, taskId);
   };
 
   render() {
